requirements.txt
python-telegram-bot==21.1.1
sqlalchemy==2.0.29
python-dotenv==1.0.1
